Factor out time_bounded_record from legacy.
