+++ lib/extensions/legacy/active_record/connection_adapters/table_definition.rb
@@ -4,14 +4,6 @@ module Extensions::Legacy::ActiveRecord::ConnectionAdapters::TableDefinition
     module_.alias_method_chain :actable, :schema_plus
   end
 
-  # Enforces a valid date range for the given table.
-  #
-  # @see ActiveRecord::Base#currently_valid? for more details.
-  def time_bounded(*args)
-    datetime :valid_from, *args
-    datetime :valid_to, *args
-  end
-
   def userstamps_with_compatibility(*args)
     userstamps_without_compatibility(false, *args)
   end
