+++ lib/extensions/time_bounded_record/active_record.rb
@@ -1,0 +1,1 @@
+module Extensions::TimeBoundedRecord::ActiveRecord; end
