+++ lib/extensions/time_bounded_record/active_record/connection_adapters.rb
@@ -1,0 +1,1 @@
+module Extensions::TimeBoundedRecord::ActiveRecord::ConnectionAdapters; end
