+++ lib/extensions/time_bounded_record/active_record/connection_adapters/table_definition.rb
@@ -1,0 +1,9 @@
+module Extensions::TimeBoundedRecord::ActiveRecord::ConnectionAdapters::TableDefinition
+  # Enforces a valid date range for the given table.
+  #
+  # @see ActiveRecord::Base#currently_valid? for more details.
+  def time_bounded(*args)
+    datetime :valid_from, *args
+    datetime :valid_to, *args
+  end
+end
